Include local graphics in a liteformat

Embed one or more local graphic files as base64 data URIs and return raw HTML that litedown::fuse() can include directly in a document. This keeps the rendered document self-contained without copying image files beside the source or output.

Usage

include_graphics(path, alt = "", class = NULL)

Arguments

path

A character vector of paths to local graphic files.

alt

Alternative text. Supply one string for all graphics or one string per path.

class

An optional CSS class string added to every image.

Value

Raw HTML output for use in a litedown code chunk.

Examples

signature <- system.file(
  "liteformats/skeletons/signature.png",
  package = "liteformats"
)
include_graphics(
  signature,
  alt = "Handwritten signature",
  class = "signature"
)
#> 
#> ``` {=html}
#> <img src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAlgAAAEsCAMAAAAo4z2kAAADbGlUWHRYTUw6Y29tLmFkb2JlLnhtcAAAAAAAPHg6eG1wbWV0YSB4bWxuczp4PSJhZG9iZTpuczptZXRhLyIgeDp4bXB0az0iWE1QIENvcmUgNi4wLjAiPgogICA8cmRmOlJERiB4bWxuczpyZGY9Imh0dHA6Ly93d3cudzMub3JnLzE5OTkvMDIvMjItcmRmLXN5bnRheC1ucyMiPgogICAgICA8cmRmOkRlc2NyaXB0aW9uIHJkZjphYm91dD0iIgogICAgICAgICAgICB4bWxuczpleGlmPSJodHRwOi8vbnMuYWRvYmUuY29tL2V4aWYvMS4wLyIKICAgICAgICAgICAgeG1sbnM6eG1wPSJodHRwOi8vbnMuYWRvYmUuY29tL3hhcC8xLjAvIgogICAgICAgICAgICB4bWxuczp0aWZmPSJodHRwOi8vbnMuYWRvYmUuY29tL3RpZmYvMS4wLyI+CiAgICAgICAgIDxleGlmOlBpeGVsWURpbWVuc2lvbj4zMDA8L2V4aWY6UGl4ZWxZRGltZW5zaW9uPgogICAgICAgICA8ZXhpZjpQaXhlbFhEaW1lbnNpb24+NjAwPC9leGlmOlBpeGVsWERpbWVuc2lvbj4KICAgICAgICAgPHhtcDpDcmVhdG9yVG9vbD5QaXhlbG1hdG9yIFBybyAzLjg8L3htcDpDcmVhdG9yVG9vbD4KICAgICAgICAgPHhtcDpNZXRhZGF0YURhdGU+MjAyNi0wNy0yOFQyMDo0OTowMC0wNzowMDwveG1wOk1ldGFkYXRhRGF0ZT4KICAgICAgICAgPHRpZmY6WFJlc29sdXRpb24+NzIwMDAwLzEwMDAwPC90aWZmOlhSZXNvbHV0aW9uPgogICAgICAgICA8dGlmZjpSZXNvbHV0aW9uVW5pdD4yPC90aWZmOlJlc29sdXRpb25Vbml0PgogICAgICAgICA8dGlmZjpZUmVzb2x1dGlvbj43MjAwMDAvMTAwMDA8L3RpZmY6WVJlc29sdXRpb24+CiAgICAgICAgIDx0aWZmOk9yaWVudGF0aW9uPjE8L3RpZmY6T3JpZW50YXRpb24+CiAgICAgIDwvcmRmOkRlc2NyaXB0aW9uPgogICA8L3JkZjpSREY+CjwveDp4bXBtZXRhPgo5rBVAAAAACXBIWXMAAAsTAAALEwEAmpwYAAADAFBMVEVHcEwAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAADR7hC8AAAA/3RSTlMACk0B/v0CBQP7+Qfq9RL2/ARUCOKA8vqmBuTwmgyX6wleD2Dd9x0NYxsOiBwsvvi35jkzEO+SbD/xTqQwIhoZ4POOC+mNKxEVeJ/aJyA+5VAYSFoTH3TYxoOimJPf7bYui9Czv86uHjJHOCgjVtWVvKvEr9PJW2H0JBfNp2ruPC03QilYRC9ZyNKW4UVnS8ze5zVkSmJtdSU7eiY2o5TUsrqH44rDOk9RiUl+XxZDyyFvd3BTrY+xx6iR6Knc2cXXbnaBFDSguMpAfEFroXnBc3FSXUzA1uxlnYJp0dt7Rn1mnrvCvaqwuVdorLWZmzFcKj3PhpClf3JVhZy0hIw2NI15AAAdm0lEQVR42uydeXwURRbHK8zkJpeQAAmEECHAEiCA3GCAcCVko1wBwy23hFsTDiPhlOUIiJxCBDmXU0BQQHcjCijIIoogArIsyIL7YUVXP7vo6vYS0lXd1dNH9Ux3Mt15v79mpmvqU/Xm2zVVr169RggEAoFAIBAIBAKBQCAQCAQCgUAgEAgEAoFAIBAIBJKR//1dJ8AKIKOVV4fjwj4BO4CM1U6uWMsevW6Xd7Pw1Z5gE5Dnmu37CCzf0IevzzmLX3aoC1YBeaxGXInaINSFfzkHrALyVIlOnqZdKN+Xf/kYmAXk8cydh4nLR8M5AAtklF7HNIUmcAAWyDDd4mEKE8YuAAvkuSrzMH2NkgAskGHywXP3EegpAAtkmNIxTG39fQEskGFagWHa+XsOwAIZpk8wTMczACyQceqFYbqcA2CBjFMrDFPuOQALZJzmYph6dwKwQMapD4ZpXiyABTJOjkAeprOnACyQgWrCwzTsMIAFMlAjeZiiLwJYIANFohsCCVeBXcEsIE9VwLkoEqwC8lgpjaVcfe0PVgF5rltSsN4Gm4AM0HkpWAVgE5ABGh0oAas92ARkhKpKwNoLJgEZoTESsHzAJCBD1IjiytcBFgEZopUUWB3LWe/7tKuQt3l3zmw/AMFoOd4Ug1WxHPW826qPQnC/a6+uASgYrO3lEqxqte5VocbqJhmAgsFqKzJvTDnp85x+LntZ4R8ACsaq/jSRebuXhx73rMrJCbZJDdZ1kXE7lYP+tmnNyesysGCo4kW2DVxh++5+H6jAFReeDzAYKR+xcfvZ3UV61skpKglgMA0sLtbenV0jYcmXejcfaDANrOK0kfbVWGG88u2f+tzo4OCGOzsKff8RPKXmgRUzzr5dfbIi6ebwdPzh2i1C52sBDqaBxW1pZteeBpBDIzG7RB/vH0T6XhkiaM0DiwubZ9Oe7sY9jKbXfwnRpO+ngQfzwOK4n2wZ5ZAYxnevebbkyjzS883Ag5lgccN/Z8OOfoV797TLpW/wpZnAg6lgcROm266fXbFroYvrtQX4mhNiaM0Fi3tpid36iVPZF8lN0Mm0PgeAMBcsLvyAvbrZkO/X4ES5q8dxt/8CQBgP1n+uUvtoN4Pt1M2tfK8Oyl4djTv9MQBhPFgVUAF1Nrq/jR4y58dnHK+YIn99I/ZkARBmgIV2+NrUV7qE79Iaheub+OtBcJ7EFLDQniDKV/qyXXpZWNKhzgoDFppInisERJgCFlrXmZrDb7bHHVyDj5M9r1SA5Pf9AYgwByzU91kb+kpr8SOw4mmc7zl4EqjJYKGUURRZl9baoJOTSvpyWLHAaZJCGogwCyzkeMZuvtJEfk2ivL2ejXv7AIgwDSyEvg2jfKVvWb2PB7X+CVEl3NmtQISJYKFm0dSglWZxX+m2km6MUimCYwCzgAgzwULJCzn7+Er9+QFYzXmCU0i/AUSYChaqVEjHlZ60cBdPMvio3uHL/AGIMBcshFKpk1JhFl4urS7pwnK1MhPLYWKUMgILVYiyia/03yUdmKxWJhWf3okAJMwGCzVcTpH1q1V9pcvVAht4EQ/pk4CE6WChExft4CudwXIelTy9cQwgYT5YyP+ftK/0r1bs4Q6W/WXiIV0JSJQCWAi1DKd8pZ9ZsIcN+IWtaqFg3MVVgESpgIXmv2R1X+mnJS2fol4Ke0hfs9nv221ObmTa7U2p1bwNLLT2fYqsI8kWs5Qfn2v0RfViM/n+tbAVVuMG4EyrSQHeBhaKT6LIil5hLUvtZ9sFxB7Sv9sIq6ZDRL7Ib70OLBTQiSKrTo6lLDWWL75avRh+euM39uEqjzocM9L7wEIol4pY5ho4LGSp44oHoClhD2kPTxoYkH7oQWbm1YJ8L0hg5x8rOdG3xAvBkkYsf+dnHUvhoaiperEcvliUu83r0zvtxzq4adMWlXU6qEqNXFJyeCNYaDQdsZzmZxlLvcGXzlcv1sqjBwqtv3GEzg7IOXVvgfmsH9pNaQGiv0VpXBk/jZIRLJRymGplYWkvMty2VAe+9H71Yu3df7ZexOe35FKbTo3TVUv1Yo/hzAsunydkfhrtG5007IIuR89S1/ZU906wkOMK1cyjpbxb666l4pxsSeyDcbntOhsWlxqtkC33pq7pGX9YuJD++EQhYXbjffbaZge5Nqetl4KF0HgqYnlyqaa/c9tS+bi01viBZ5Gf62qXX260ch5mPaFGZMSkTgnn1BRVV+U6679rRD+Z1kz2WrBQsy3UUK/bmTum3mvHjo641Pn5rB06HfjuW+oQLp2iURB7SPP0tGvfNU5FFXX8Ge6SSYSaOERSYT3Gyi7LtSbLe8FCPamI5eF66HBs7zJI/FikA7pmIO5bKpX1eS53+HJXdPx9/eTkVLWUva4DrmAl15bWV5PtoHbcU3KNmejFYEkiltnvgeQWTaT91IOWB5YiZ5z7Mrol2Oe4M0ZwGurH/jtMdAFrf23XCgcw1XVQ/JWQUWWTB1MfWA8bLb5J9zB6oDKj5My+pUCV4YR20w2wFJn0z2X8NTaxWi6jCacp9syIk6Rg3X9Fpj7fuizmrij6RutZDcpme10vWOgLESTT0lm+seR9JbtnKQzt/n8ccK3YL1Q02mNLjcff0zoOkqMzRdaaQG2uuH8w/w6VJWCN7Shb4QaGqnaIQ+hWkCcHtvRysFCz5kK7n62k/e/ZVsXwr+yQ+cbaBu+SbZsITy2FExhxWs+6zNCVIuuJOxyLqjL/DE4arMcVsH2coa6pojMwTUnYEOvfS9mBhTa01jHP6VOkbvoGtFe5e8b1keLLzTy1VDP8zV6M632mFFl9azNxxW1jtekC+vEYu8WV9LjbgbzeqV1Vd+Hpsb5Pi/wovbweLPSc6N9Qw+0zvYeW7SdHoOC5ba5GLmo79e6RmnJDv0eWIv6hnxGjh5Rh5RU/wbUfozb0TZg+tNf5jRQTrCYdLwYr4juhhqgP1yNUVY9T+Kzw5cyHb/fiN/W9Hyy0TpgANFddbTWtqX1XTxjkq3gtJs5TSz2BS2tmoOjMnCLLf5JrS2PJyJsjuu2Yt7TPiMBKuSdUcOpRgnABrN3aVQnfblQ8+O7j33yJLAAWyggRjt+r7O10jeI801WPLeVQye9OaxlziizXbUsqjDBbFM3NuqWdJYA1403yujHvERXAekyzpmDy07ROKH6PlzojLAEWWiXYTnlx1n2QZ1iFZxpgqcV88fe0Ct5h3Yi5ofVE6ZnChf2M9nyBfGOYcJqzzj4kBUt7lJ5PN8qB52fHrAEWEuKVY5QmJdWmeELV4Htr6hthKZy3tr9WwYGMKbJyNJ9Uflufe6BYzWUsUOULJAWLYck6jExS4ovfbte7HVTWYCXGCAs7hSKn3GXKubF67jiHQZa6wvqbHGQLjm8VpMUV+li41I6tkXvlxmthubGNYx12EXqeDH3U7aK58+AtYKF6gitcPslRnltQxdyNHBtvpKXwLnSgVgRZb6aNquTBmlwh0UqYMbprlowlhE1Dv2nMKxBUnx4sA7BLcCGyCliiP0PZI1MJuifuzn5plxv6GW2pbNbpSQZTiqzJ2lwh0VL4hN67VM5D+AP5cLZmTbm46OJHQ34b/PYz64CVSPykMXL+9191QdXx9m8FoaZYyhHFFvROzomppsgqcGn5b66FRF5zxtCinS7VXooTrl4gq0Tt6t7DZTsVv6uBJ6TOROuAhSLJVw+5XpwnMVR0do2e2Qtm9T7z1YgO4VIr/mlPNfMsNYVx366aUztFVkoHadP/J3PHifbZGdvoskEUJJ6cHcOfFmlW5Ef+qR9FwpKYiSnIQmAlE4+4a57r0M6SSCJxcIF/+32v/hKbdfij/vzVc2Zaijke5l3tFFmRUgBelykk2tq8y9jGS9J6z8vdGtq+ODSGRCUVj++nyQ7kPCuBJayCG8dLLw2QTMnlswNN1w6I89xSOEhwGauHVDlFVrzUK/BfuY3FZ3SfjAmQrjSLxPUGTGN3j7bERYszCviQHc1rDkuB1ZRTeoBk+yqUnUJmyVeAQw/Gm2kp/KjCxlpB+lM1U2SdkUbdy57LGqI7ljhUuoyh3F9ddbhHyb1+9uEbIQlVG2QpsMg97rIupF1YQWMVvv837ZxoBlhqMaNPaaBWiqy6ElfDZHkPhih6rKt7YNEej1o63KO4r9xQhD4k9S1EFgOLdFni1k6ns2opRkBs5kukm2qpRYyprz7QKvcW/fO/Iz8E9hVKfOlwC6wQemRqwe4eJeuGKD/0QKhwjtXACsa3cAh97x6lxnXluKkX+SCrAFMttZL/0kBGD6ni5id9JOeewkr2Z9FBJuQWWJIQhiJ29yhZizcS3YWlfVTVALCEZXJD8af51OEVlUDPFzSn1UZYqlIY227hvzRSZLWjDzorTdm2CmVYfW30jk5NOiQiIITdPUq2JbqI4nb7+VgPrEzZs5lTxWZSiUyfwXsS00y2FP8E+/BQNg+pUoqst6kQPsVjDRcl8a8MWkeBdV1+8cHiHiXRtzGiGW7Z5Ov1DCyyLhRnnzopttJOhklNnsmWWup6FlTNQ6qQIstPfKTDuU6xGsGDFxWgs4ElPZX4bvZwrPEZqJtcyOQaZEGwuuEv3xB9KM4AGKv27T9r7tQaYyk8URvF5iFViCdeIW7DKWWflNDkJNYGthBX/YvioK1ZzyEZa/2fvSuPiuJI4+0DhksQBAUFhKCAYhA5VNSIx/PciIrBeCvifSGbyHqxRmMU0XggxlWJMckTJbpe2U28olFj4kHQbHQ3MYnxNqs8TaJrfIkvO6swXV3ddXTNTDfKvu/3F2+mqqb49ddVX331HTdcaqJgufjh9y1V2Kxr6NHoNANNZ8qWg8kngt9sJ9efWGXzjWPv70qj06LzwzM0+WhDtgeJm0eHkmzlNJBqomAhm80VKks/cFPSfKb/JhrE1DYx69fLHH/i8aqTLsctOlNptU50fl05a7xHHXHzKBk7tCJGqpmCJTsfV9A0rBxucodYH/2IP4OYSg0SSiwk70g/v9CoouUPN/KCH+GXSoXOskYdP3qcPchEJVRN9BXoiQ+dqflynrh5dBbB1vPbpRoqWPLrdAd9olibssO4XWVXkcPsJh2MYsq23NT2ELKQRqkcek5tXZWiDY4czBzDXcmbIhy+cBIbmbhcnyFsHnUZRHiNDJBqqGAlk95uyBPOj3+fkempb/ZbZhRTsofmVm6rfGH3sS3MMTCbxCeis2uDjUykpH5L1DzqMk07y+lPMie/c4KF7keL0EcoEVkFt6dFfrPr9NA1QBnA1D2RyOS7woLFdID4Emu0Q3RyWNberwjtYbGgedRyRTvJXF/B37f07v20CdZFIoA9Go3HL3f0ZwGDRISPYUzJvr8pvG5ponLVlzXCmzOxIBtBM25q2QqOecY/Ssw86r9PO8kWotronxKs1l+eMsFC4YXz5E9QrLhbJJdN2a85hJMQrZ1xTEXYbjWvilhIrdmB/FQyexgDTP4Ka7RXaK5JH6lMdSnMQybfPOq9RzvHDwIE2aqK00w2T7DOOdAb+Q2j/QxVseBXVx4ncp3Wy0Cm5Bu8SbyeckqqtZJLj1mT2tUrZAjWJYYhdqnonQNa8TXB+g0tzFeXax71JXKVdxFN4W27gl1knmD1sb9zpLxOK37ie4Vs7igD82Ge990mA5lKsy0MowUspMjetZYhWPR0JB3VDsa6EdUe+fe0A5MOGNNEfAYXHNQmaDsi+gh7CSkuTgnWPPs7NyUc/byRk+0MTr8i9KMTecOfN5KpUNvnPwqso/LFuQXbpvDNkRrUlaZ5ujp+hRGfryQltpRoliuQGWt1W80wPv1E2fqSZT4zULAcSHODLvJRurxEEZepreghjRayhBvCVIZN/5kToG8hlf2oXsc2KczZejh11dDk5Xbj7twZ26h5eI4xdXeOYL2ovVB1TRRla0F9odk6J1j2e+10l5Xdf1FMMuwT7ARkxfbi5+i8ZihTlwlbLoG/qE99LkpspFsK5mNGywM+QSsoTTizHLDfh7rFHiJaPqebyy99NDHMTlG2YmU3AOtK806F8XZ39ThB3rYiBy12XvUphbpZH2zYaChTllzddz9fHRnfGjd2Yk4zR8meU2dqny77UJjZnnXm/IytbLBm/RoZ8q9Hq7Ic/tsq7jnhsGAV2t31e/Q6R6PPUOBdCKtXH+XVbqETN1NoLFOf2raMOpl6BvrKp+h/X3lUWR5SFscx2tKcfLqsSP2TLdlmDLKm4hK+YB27ge+AVv2gJ6ra+UjVmG2eYGXb2zMT6emYn1N7vbW1SNFVBvbQ+QUvg5mSLdwXIvUspEMlqXin2rykBHW5arv3OK60HBvFu3iw/P0EsW9h+UbJOj4neILlUYS7RURxjgA0hIVy47kNEyx7g7B7oo0qBLvgQB6fG+m7537MtXuK3n9uNFO+8mJ5nLFPu8svfcNbefHYM+umOqJqNaG72PGuXLbE0vytvS8TGVnHtkNX349WfnJaczmCtXyjyvH0JvrzWxG2FmLZApZmGC9YEfLgb9vX71kleTsWLmVBzkO5tE7J7yj/TLyut9J4w5lC2TXuMTI/jKDuUKGPXQuVYFV1caiID5VDWVAbxUONKCKRfPqAduDvSh+NXG7lbBqtmIKVpLaJ3pvyo4DvhaKM3sYdOBZJJkAmotyuXmuXUpWJXVyfkV8DMYUgX/c3TGDqPyhgn14Vg1rq4C2L2vSgcjNIPKQ0LHwc8C3Xq8hRj9zsv0Rap+y6lb6wSmLJDeSEwumC5dJ/pLqMT2iAlIGOBHN02ZoYiBtGzImTvmZfmgGbBq68fJ1xT5bVHN01TlV65pT+j5jBlJJE9gzFhJHqR8naZTsDRrpRvBbSG2E2pI8rVQK5QlorlRXkSpB23KUzPDSLPO16azhNsGLaHFKP1bYyTOSqgON01dNTJfnyNCmJ5BL7QxR64mmDVRb2r5mCldxa9dCE6uCYwJQ7thlnjyrD1XD3zDZdSbnyQjcz04mL4rB6uJFhSQO1KIxHI6/bQ1QJC/+H7B+DuX1SilxPJwQrtvSO1tA+rsq2fQt9cCGds7R/rlLNrK12mCNX0pSq00RQnGgHl8tzmeGopYpbsurz7V+o7Ic+YrlizGAqRlX4whofOOfMng1DhgzJy/KibYNtXyWPnDbBSj6C60y1+2mPL4N3fLogLmnd5XLSv7rzJ7FoWCzpEcXbbJBKsDKSjo4hlr6G8i/PVlbPApaD0aKhmho9B82riv6g0naeIth69tHnOWHOk2gutt51R6stgp0EU76awlS6PTnCfdaRZvlKwbK8HqxyYH55u9amwUZUl540a6B1BGW2imEua28nmuh7dVGcibBULLkkC+6rjgRr/3e31mGmyZVkyb99fqJYSFWt5ePUL4w2fH6A1gHEvcPFRnM1/82YnqJTM4Wp3rWF5UqVJgddKnxR1PyPgep1DUtP8YzOmPU/VF/KKjncr3FMbwy4bUjDj9GYicR1z7dJHcMeP9aA1HkppU1HlUSR3QvmS9WM4iU38vZ/X9RuUgfZtFKr+Oy0OZqrCFciLYOL8tByRm0qPxNICTgdJl6m3Bymxp8QlCs/VQrMF5ntWozHlUkf3pCFvbTW1QJFSaPMtZQ7wZaa0+8o4uRRv2sdZu+E3e7VLVco6/7j3WB4bkl2YTxlYk0o22ew3sOa29ieeZjDVIOhYoKlzmfEEqyo3erlnjP4wWXE1bi/sr/RMg1E+rFXq3GE02JMlvgu33BNgFT96Ks/MddpNPN1Yx314j37XGDNYupcgcCAGpMxXbDc8tZqxo4OZ6x/x6dSos8WKA3KaDP9mDG5+FCa+0hqoCBZLxV5S08CF3Rn1ol+9kr1420t5+1OCm0WU5YXBuqNGKSRmK20t6uCcpheSONg5276PW9dpQm1wRvUuQ18n3Fr3F2Eryb1pkhPCOU6UwsfzApsaM1ZAbo7MBPzmIrrVWLbVV1Dho+9XnFn0+lbN7tgxwVthzlkUAX9eNtHM+egkkvF3DM5zx/kHVL9f7iZY3x5qMNVSOgrLtITw4QQ3twKmrI3nR6MjcD1+l3HpmIqU7HFr6yalaHsUBGYzemhtvFJzWlhJLPMdPIDxTSeff8c76CqXEEy0s/FjFTvFMMu6kTul7Vvy1YX9k0NkJ4o5q9gzc3zbX6exsvUJeXBLsfnUo1MvcsNoz+JXfcduM2N95EW1W0zeP3ZslkROr+YgkZkhfZ5NLU58PgVnF8mZMb2PldOVhNznd7opzTpKUD390gN3nNx8/66wX1vaMTA68y2RCcX3+piKg630lFO/8cuHXR7dAJpUjJqs8Wo3zy7t5KvwG2cY03HVVvaJc6yx3fcfWu9JR+8dKHqTQjPzftmdbL01MBjfv760KuHEzp7evXNyskbdqq/mJl2gOzQEdS3RaO/LTQmA2Z1MOWiOoK1pM8jzdfw3222YK1ZaUJrpaX5ukv/N3BPji7edcy7pjGlzkFSWwIAjEC6ZruNBkoARuAjjRLXDygBGAAiO9bvwAnAeUQT586WQArAaViQ43QhaO8A44CCsA43CwLtHWAUnpO9V2b2lkpAewcYhIhsWZj+iiXvGgzEAJzDMPzC7iRo7wBjgJL+r3wcPeMdBdo7wAj4ytftnlVBrP8E7R1gBPZpktb+Dto7wAAgk/s1m8dff9DeAc7jWTnuMUSO4op1A+0d4CxckFP5WfRZFmjvAGexhpKy9jxo7wAnMVmOf03AcrXtBu0d4Bz8UaZRvOJcGWjvAOdQjxo64Qu2d4BTeFNOc9JJnS+gIWjvACdQ6xCj7NJI0N4BTgCVxQ3WfPEuaO8Ax3FRFp862oUpH7R3gMOIRKHe32i/QrWuRwJNAHvRDWXGJuKQUYkK0N4B9qIY+bbXJb8sBO0d4CBG0zOHV+EEaO8Ax4C8Rj1/pnw7DrR3gEPwR+ETXWhf7wftHeAQkFPDgVja16haQAJQBbADviOweC8KkCHLC7gC2AFUa3cFPdkgKuJr9QeyAMIIQxlAttAboCK+1lhgCyAM5Mn3HSM7qjsSrHRgCyAKj5W6ZcxRJZIOQBdAFL+iCg/M5MetBGtwAgAKFstS8wyziezqZ50EdAEEsQ7VNWAf+VC5z4XAF0AQ6L7ma3YbVHasFPgCiMFXrs7Vl1OVZpD+bgkAqHCTe0toQw6vPCUAQMEfRPRyVGO2FxAGEEOCTWQCeaWjUGqj34AwgBAWieyEyrLWDRgDCOF9IQvVBrlVa2AMIIQxIjuhVCEL1n1gDCCCsCCRnVCJhX4AlAFE0F/srgYluD0ClAFEUCRHDPLLCRfIgvUTUAYQwXqbwBznNwuXBWs5UAYQwX2bwDTntvKXQ6GtiUAZQATBNoF5jduqGfIgTQPKACK4bhOYydxWSbJchQBjACFMt0lMDLfVVFmwzgBjACF0rRIYP36r32TB2g+MAURgsVWdGMFvhjKGNAbKAP9r7+5ZowgCMAAXdwmKip4oQdAmBqKVjWgTDBJMAhG0EGwMgvgTBA8UxEKw0cLGwqBFbELs/CCFkXSKVRTBRhBLgwiKaURYQXb2TO4yN7uX8nnK3aleXnaXnd2ZFEfDBnLxYXtCsUZERoot+XuER9FRE6FXR/aJjCT5lkxnooNuWWyGkvLfbwaig46FYt0RGGkW8sqMRsbUZvNB9VGBkeZ63plnkTE3wwVrQV4kCiuuvYiMmfMnBWWFyZpPkTvhUJY07wMtjR1d31DtDr16Iy6SXe42WVML04nR5zBY62nemo9j3S5Yr/qlRbJzYU21vd0uWIeERQkv8940T3Q8fTVLWDME2nzeaJfCf8bDw322LCrK6F8M1VlpP3lgqViVbUxUlHK22B1gpu1c+CQ+6/spKEp6Hdqz/d66M/PFTxSPxURZD4rFtptrdhlvfC96dWWbmCjtV1GgrRdaHzC8P18cHvDlKBU0ipthlu1c/ToyPPnj7upU61j2TkZUMXwpi7kvIao5PBvplRVmqGx8eqNa1b9Ih+omvnXu1a4nsqEXtbd9HXq1ZPMcenXjYn1drU7+9v6KTXD6YfO/Wg0tD4qEzTH44fip59embv/Zf9CelwAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAADQg7+EoYmUm78u8AAAAABJRU5ErkJggg==" alt="Handwritten signature" class="signature" />
#> ```